Add cache for params parsed from on-disk ics files
