+++ web/views/page_view.ex
@@ -1,4 +1,6 @@
 defmodule Lonestarelixir.PageView do
+  alias Lonestarelixir.GoogleCalendarParamsCache
+
   use Lonestarelixir.Web, :view
 
   # Constants
@@ -14,7 +16,7 @@ defmodule Lonestarelixir.PageView do
         end,
         content_tag(:td) do
           ics_path
-          |> google_calendar_params_from_ics_path()
+          |> GoogleCalendarParamsCache.from()
           |> google_calendar_link()
         end
       ]
@@ -38,57 +40,9 @@ defmodule Lonestarelixir.PageView do
     end
   end
 
-  def google_calendar_params_from_ics_path(ics_path) do
-    "web/static/assets#{ics_path}.ics"
-    |> File.stream!()
-    |> Enum.reduce(
-         %{},
-         fn
-           "BEGIN:VALARM\n", acc ->
-             {:alarm, acc}
-           "END:VALARM\n", {:alarm, acc} ->
-             acc
-           _, {:alarm, acc} ->
-             {:alarm, acc}
-           "DTSTART:" <> date_start, acc ->
-             Map.put(acc, :dates, String.trim_trailing(date_start))
-           "DTEND:" <> date_end, acc = %{dates: date_start} ->
-             %{acc | dates: "#{date_start}/#{String.trim_trailing(date_end)}"}
-           "DESCRIPTION:" <> description_line, acc ->
-             {:details, Map.put(acc, :details, [trim_ics_line(description_line)])}
-           " " <> description_line, {:details, acc = %{details: description_lines}} ->
-             {:details, %{acc | details: [trim_ics_line(description_line) | description_lines]}}
-           _, {:details, acc = %{details: description_lines}} ->
-             %{acc | details: ics_lines_to_string(description_lines)}
-           "SUMMARY:" <> summary_line, acc ->
-             {:text, Map.put(acc, :text, [trim_ics_line(summary_line)])}
-           " " <> summary_line, {:text, acc = %{text: summary_lines}} ->
-             {:text, %{acc | text: [trim_ics_line(summary_line) | summary_lines]}}
-           _, {:text, acc = %{text: summary_lines}} ->
-             %{acc | text: ics_lines_to_string(summary_lines)}
-           _, acc ->
-             acc
-         end
-       )
-  end
-
-  def ics_link(path, opts \\ []) do
-    link(to: "#{path}.ics", title: Keyword.get(opts, :title, "Add session to iCal/Outlook")) do
+  def ics_link(path) do
+    link(to: "#{path}.ics", title: "Add session to iCal/Outlook") do
       tag(:img,  height: "32px", src: "/images/ical.png", width: "32px")
     end
-  end
-
-  ## Private Functions
-
-  defp trim_ics_line(ics_line), do: String.trim_trailing(ics_line, "\n")
-
-
-  defp ics_lines_to_string(ics_lines) do
-    ics_lines
-    |> Enum.reverse()
-    |> Enum.join()
-    |> String.trim_trailing()
-    |> String.replace(~r/\\,/, ",")
-    |> String.replace("\\n", "\n")
   end
 end
